chore: 🔧 rename calendarStore to meetingSetupStore
